feat(tech): protect API route exposed on server
By preventing them to be called outside browser navigation

We use a session UUID, valid during the entire session TTL. By default,
it's 14 days. If the cookie isn't present in the request, the protected
route call throw a 401. Every HTML route in /page will start a session
if none exists.

 ###Drawbacks

This solution is imperfect, as it's straightforward to bypass : you just
need to call a HTML page (like /), get the cookie, and use it in your
API request without limit during the next 14 days.

The problem is that we cannot shorten the TTL of the session, since it's
the same object used for auth as an agent.

 ###Futur improvement

We might want to couple this with rate limiting. We can use the session
ID for as a key for rate limiting. However, it can be easy to overcome
with a pool of session call sequentially... Maybe the solution is to use
IP as well ?

fix #649 (partially)