test: add snapshot testing for API clients
- Add a new command `test:client-api` to run snapshot tests for API
  clients
- Remove old mocks
- Use snapshots for E2E testing

We need to be sure that the API clients response will not change without
us knowing it. As we are not the owner of the API, we can't be sure that
breaking changes will occurs silently in the future.

By introducing snapshot regression testing, we can check periodically
that the API client response is not changing.

Besides, the snapshot are used to mock the API client in E2E testing.
The means that we can test the whole application without having to mock
the API client manually.

The test interface is splitted coherently and allows for full chain test
of the application.

+------+
| APIs |
+------+
    |
    v
+------+
|Client|
+------+
    |
    |  <-- Snapshot testing
    v
+------------+
|Application |
+------------+
    |
    |  <-- E2E testing
    v
+------+
| User |
+------+

fix #643