feat: change insee and rne sections order
In `Dirigeants` section, we display the `Avis de situation` section
before the `RNE` section. The reason is that the inpi often takes very
long to load. In order to improve UX, we display the `Avis de situation`
which is already loaded. This means that the UI changes happens outside
of the viewport.