ci: run external API on pull request
We want to check error regularly, and to really look at them when they
trigger. Besides, some changes in the code can lead to regressions in
the client test, and we want to be aware of them as soon as possible.

In return, we only trigger the test once per day, as it is not necessary
to spam the tchap channel when it's failing consecutively.