feat: add error boundary for client errors
This commit adds an error boundary to catch client errors and display a
message to the user.

It allows to pass additionnal metadata to Sentry (like the component
stack), and is the [recommanded
way](https://nextjs.org/docs/pages/building-your-application/configuring/error-handling)
to handle clients errors in Next.js.

Besides, it changes the error message for server errors to be less scary
(remove the "⚠️").